% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euRopeana-package.R
\docType{package}
\name{euRopeana-package}
\alias{euRopeana}
\alias{euRopeana-package}
\title{euRopeana: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
