% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_key}
\alias{get_key}
\title{Get API key}
\usage{
get_key()
}
\value{
character with the API key stored as environmental variable
}
\description{
This function is a simple wrapper of the `Sys.getenv` base
function. It gets the value of the environmental variable `EUROPEANA_KEY`.
}
\seealso{
\code{\link{set_key}}
}
