% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key.R
\name{set_key}
\alias{set_key}
\title{Set API key}
\usage{
set_key(api_key)
}
\arguments{
\item{api_key}{string with the API key}
}
\value{
No return value, called for setting the environmental variable
`EUROPEANA_KEY`.
}
\description{
This function is a simple wrapper of the `Sys.setenv` base
function. It sets the value of the environmental variable `EUROPEANA_KEY`.
Alternatively, use .Renviron to set the key. Get and API key in the
following link https://pro.europeana.eu/page/get-api.
}
\seealso{
\code{\link{setkey}}
}
