% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_search_items.R
\name{tidy_search_items}
\alias{tidy_search_items}
\title{Tidy search API response}
\usage{
tidy_search_items(resp)
}
\arguments{
\item{resp, }{and S3 object of type `europeana_search_api`.}
}
\description{
Transforms API response to a tidy `data.table` for easier manipulation
}
