% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_media.R
\name{download_media}
\alias{download_media}
\title{Download Media}
\usage{
download_media(urls, dest_dir = NULL, quiet = TRUE)
}
\arguments{
\item{urls, }{character vector of fiel urls}

\item{dest_dir, }{destination directory}

\item{quiet, }{boolean to suppress download file messages}
}
\value{
destination folder
}
\description{
Function that downloads files from the given urls. The query
parameter `value` is used as the file name.
}
