% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cursored_search.R
\name{tidy_cursored_search}
\alias{tidy_cursored_search}
\title{Tidy Cursored Search}
\usage{
tidy_cursored_search(query, rows = 100, max_items = 10000, ...)
}
\arguments{
\item{query, }{character string with the search term(s)}

\item{rows, }{numeric that indicates the number of records to return in each
batch with limit up to 100.}

\item{max_items, }{numeric that indicates max items collected}

\item{..., }{other params passed to get request, see also `query_search_api()`}
}
\description{
This function is a "runner" of a particular query that consequently makes
API requests until maximum items are reached or all related items have been
collected.
}
